+++ contracts/BMath.sol
@@ -26,20 +26,17 @@ contract BMath is BConst, BNum {
     // sF = swapFee                                                                              //
     **********************************************************************************************/
     function calcSpotPrice(
-        uint tokenBalanceIn,
-        uint tokenWeightIn,
-        uint tokenBalanceOut,
-        uint tokenWeightOut,
-        uint swapFee
-    )
-        public pure
-        returns (uint spotPrice)
-    {
-        uint numer = bdiv(tokenBalanceIn, tokenWeightIn);
-        uint denom = bdiv(tokenBalanceOut, tokenWeightOut);
-        uint ratio = bdiv(numer, denom);
-        uint scale = bdiv(BONE, bsub(BONE, swapFee));
-        return  (spotPrice = bmul(ratio, scale));
+        uint256 tokenBalanceIn,
+        uint256 tokenWeightIn,
+        uint256 tokenBalanceOut,
+        uint256 tokenWeightOut,
+        uint256 swapFee
+    ) public pure returns (uint256 spotPrice) {
+        uint256 numer = bdiv(tokenBalanceIn, tokenWeightIn);
+        uint256 denom = bdiv(tokenBalanceOut, tokenWeightOut);
+        uint256 ratio = bdiv(numer, denom);
+        uint256 scale = bdiv(BONE, bsub(BONE, swapFee));
+        return (spotPrice = bmul(ratio, scale));
     }
 
     /**********************************************************************************************
@@ -53,22 +50,19 @@ contract BMath is BConst, BNum {
     // sF = swapFee                                                                              //
     **********************************************************************************************/
     function calcOutGivenIn(
-        uint tokenBalanceIn,
-        uint tokenWeightIn,
-        uint tokenBalanceOut,
-        uint tokenWeightOut,
-        uint tokenAmountIn,
-        uint swapFee
-    )
-        public pure
-        returns (uint tokenAmountOut)
-    {
-        uint weightRatio = bdiv(tokenWeightIn, tokenWeightOut);
-        uint adjustedIn = bsub(BONE, swapFee);
+        uint256 tokenBalanceIn,
+        uint256 tokenWeightIn,
+        uint256 tokenBalanceOut,
+        uint256 tokenWeightOut,
+        uint256 tokenAmountIn,
+        uint256 swapFee
+    ) public pure returns (uint256 tokenAmountOut) {
+        uint256 weightRatio = bdiv(tokenWeightIn, tokenWeightOut);
+        uint256 adjustedIn = bsub(BONE, swapFee);
         adjustedIn = bmul(tokenAmountIn, adjustedIn);
-        uint y = bdiv(tokenBalanceIn, badd(tokenBalanceIn, adjustedIn));
-        uint foo = bpow(y, weightRatio);
-        uint bar = bsub(BONE, foo);
+        uint256 y = bdiv(tokenBalanceIn, badd(tokenBalanceIn, adjustedIn));
+        uint256 foo = bpow(y, weightRatio);
+        uint256 bar = bsub(BONE, foo);
         tokenAmountOut = bmul(tokenBalanceOut, bar);
         return tokenAmountOut;
     }
@@ -84,20 +78,17 @@ contract BMath is BConst, BNum {
     // sF = swapFee                                                                              //
     **********************************************************************************************/
     function calcInGivenOut(
-        uint tokenBalanceIn,
-        uint tokenWeightIn,
-        uint tokenBalanceOut,
-        uint tokenWeightOut,
-        uint tokenAmountOut,
-        uint swapFee
-    )
-        public pure
-        returns (uint tokenAmountIn)
-    {
-        uint weightRatio = bdiv(tokenWeightOut, tokenWeightIn);
-        uint diff = bsub(tokenBalanceOut, tokenAmountOut);
-        uint y = bdiv(tokenBalanceOut, diff);
-        uint foo = bpow(y, weightRatio);
+        uint256 tokenBalanceIn,
+        uint256 tokenWeightIn,
+        uint256 tokenBalanceOut,
+        uint256 tokenWeightOut,
+        uint256 tokenAmountOut,
+        uint256 swapFee
+    ) public pure returns (uint256 tokenAmountIn) {
+        uint256 weightRatio = bdiv(tokenWeightOut, tokenWeightIn);
+        uint256 diff = bsub(tokenBalanceOut, tokenAmountOut);
+        uint256 y = bdiv(tokenBalanceOut, diff);
+        uint256 foo = bpow(y, weightRatio);
         foo = bsub(foo, BONE);
         tokenAmountIn = bsub(BONE, swapFee);
         tokenAmountIn = bdiv(bmul(tokenBalanceIn, foo), tokenAmountIn);
@@ -115,30 +106,27 @@ contract BMath is BConst, BNum {
     // sF = swapFee                \                                              /              //
     **********************************************************************************************/
     function calcPoolOutGivenSingleIn(
-        uint tokenBalanceIn,
-        uint tokenWeightIn,
-        uint poolSupply,
-        uint totalWeight,
-        uint tokenAmountIn,
-        uint swapFee
-    )
-        public pure
-        returns (uint poolAmountOut)
-    {
+        uint256 tokenBalanceIn,
+        uint256 tokenWeightIn,
+        uint256 poolSupply,
+        uint256 totalWeight,
+        uint256 tokenAmountIn,
+        uint256 swapFee
+    ) public pure returns (uint256 poolAmountOut) {
         // Charge the trading fee for the proportion of tokenAi
         ///  which is implicitly traded to the other pool tokens.
         // That proportion is (1- weightTokenIn)
         // tokenAiAfterFee = tAi * (1 - (1-weightTi) * poolFee);
-        uint normalizedWeight = bdiv(tokenWeightIn, totalWeight);
-        uint zaz = bmul(bsub(BONE, normalizedWeight), swapFee);
-        uint tokenAmountInAfterFee = bmul(tokenAmountIn, bsub(BONE, zaz));
+        uint256 normalizedWeight = bdiv(tokenWeightIn, totalWeight);
+        uint256 zaz = bmul(bsub(BONE, normalizedWeight), swapFee);
+        uint256 tokenAmountInAfterFee = bmul(tokenAmountIn, bsub(BONE, zaz));
 
-        uint newTokenBalanceIn = badd(tokenBalanceIn, tokenAmountInAfterFee);
-        uint tokenInRatio = bdiv(newTokenBalanceIn, tokenBalanceIn);
+        uint256 newTokenBalanceIn = badd(tokenBalanceIn, tokenAmountInAfterFee);
+        uint256 tokenInRatio = bdiv(newTokenBalanceIn, tokenBalanceIn);
 
         // uint newPoolSupply = (ratioTi ^ weightTi) * poolSupply;
-        uint poolRatio = bpow(tokenInRatio, normalizedWeight);
-        uint newPoolSupply = bmul(poolRatio, poolSupply);
+        uint256 poolRatio = bpow(tokenInRatio, normalizedWeight);
+        uint256 newPoolSupply = bmul(poolRatio, poolSupply);
         poolAmountOut = bsub(newPoolSupply, poolSupply);
         return poolAmountOut;
     }
@@ -154,29 +142,26 @@ contract BMath is BConst, BNum {
     // sF = swapFee                               \      tW  /                                   //
     **********************************************************************************************/
     function calcSingleInGivenPoolOut(
-        uint tokenBalanceIn,
-        uint tokenWeightIn,
-        uint poolSupply,
-        uint totalWeight,
-        uint poolAmountOut,
-        uint swapFee
-    )
-        public pure
-        returns (uint tokenAmountIn)
-    {
-        uint normalizedWeight = bdiv(tokenWeightIn, totalWeight);
-        uint newPoolSupply = badd(poolSupply, poolAmountOut);
-        uint poolRatio = bdiv(newPoolSupply, poolSupply);
+        uint256 tokenBalanceIn,
+        uint256 tokenWeightIn,
+        uint256 poolSupply,
+        uint256 totalWeight,
+        uint256 poolAmountOut,
+        uint256 swapFee
+    ) public pure returns (uint256 tokenAmountIn) {
+        uint256 normalizedWeight = bdiv(tokenWeightIn, totalWeight);
+        uint256 newPoolSupply = badd(poolSupply, poolAmountOut);
+        uint256 poolRatio = bdiv(newPoolSupply, poolSupply);
 
         //uint newBalTi = poolRatio^(1/weightTi) * balTi;
-        uint boo = bdiv(BONE, normalizedWeight);
-        uint tokenInRatio = bpow(poolRatio, boo);
-        uint newTokenBalanceIn = bmul(tokenInRatio, tokenBalanceIn);
-        uint tokenAmountInAfterFee = bsub(newTokenBalanceIn, tokenBalanceIn);
+        uint256 boo = bdiv(BONE, normalizedWeight);
+        uint256 tokenInRatio = bpow(poolRatio, boo);
+        uint256 newTokenBalanceIn = bmul(tokenInRatio, tokenBalanceIn);
+        uint256 tokenAmountInAfterFee = bsub(newTokenBalanceIn, tokenBalanceIn);
         // Do reverse order of fees charged in joinswap_ExternAmountIn, this way
         //     ``` pAo == joinswap_ExternAmountIn(Ti, joinswap_PoolAmountOut(pAo, Ti)) ```
         //uint tAi = tAiAfterFee / (1 - (1-weightTi) * swapFee) ;
-        uint zar = bmul(bsub(BONE, normalizedWeight), swapFee);
+        uint256 zar = bmul(bsub(BONE, normalizedWeight), swapFee);
         tokenAmountIn = bdiv(tokenAmountInAfterFee, bsub(BONE, zar));
         return tokenAmountIn;
     }
@@ -193,32 +178,35 @@ contract BMath is BConst, BNum {
     // eF = exitFee                        \     \      tW /       /                             //
     **********************************************************************************************/
     function calcSingleOutGivenPoolIn(
-        uint tokenBalanceOut,
-        uint tokenWeightOut,
-        uint poolSupply,
-        uint totalWeight,
-        uint poolAmountIn,
-        uint swapFee
-    )
-        public pure
-        returns (uint tokenAmountOut)
-    {
-        uint normalizedWeight = bdiv(tokenWeightOut, totalWeight);
+        uint256 tokenBalanceOut,
+        uint256 tokenWeightOut,
+        uint256 poolSupply,
+        uint256 totalWeight,
+        uint256 poolAmountIn,
+        uint256 swapFee
+    ) public pure returns (uint256 tokenAmountOut) {
+        uint256 normalizedWeight = bdiv(tokenWeightOut, totalWeight);
         // charge exit fee on the pool token side
         // pAiAfterExitFee = pAi*(1-exitFee)
-        uint poolAmountInAfterExitFee = bmul(poolAmountIn, bsub(BONE, EXIT_FEE));
-        uint newPoolSupply = bsub(poolSupply, poolAmountInAfterExitFee);
-        uint poolRatio = bdiv(newPoolSupply, poolSupply);
+        uint256 poolAmountInAfterExitFee = bmul(
+            poolAmountIn,
+            bsub(BONE, EXIT_FEE)
+        );
+        uint256 newPoolSupply = bsub(poolSupply, poolAmountInAfterExitFee);
+        uint256 poolRatio = bdiv(newPoolSupply, poolSupply);
 
         // newBalTo = poolRatio^(1/weightTo) * balTo;
-        uint tokenOutRatio = bpow(poolRatio, bdiv(BONE, normalizedWeight));
-        uint newTokenBalanceOut = bmul(tokenOutRatio, tokenBalanceOut);
+        uint256 tokenOutRatio = bpow(poolRatio, bdiv(BONE, normalizedWeight));
+        uint256 newTokenBalanceOut = bmul(tokenOutRatio, tokenBalanceOut);
 
-        uint tokenAmountOutBeforeSwapFee = bsub(tokenBalanceOut, newTokenBalanceOut);
+        uint256 tokenAmountOutBeforeSwapFee = bsub(
+            tokenBalanceOut,
+            newTokenBalanceOut
+        );
 
         // charge swap fee on the output token side
         //uint tAo = tAoBeforeSwapFee * (1 - (1-weightTo) * swapFee)
-        uint zaz = bmul(bsub(BONE, normalizedWeight), swapFee);
+        uint256 zaz = bmul(bsub(BONE, normalizedWeight), swapFee);
         tokenAmountOut = bmul(tokenAmountOutBeforeSwapFee, bsub(BONE, zaz));
         return tokenAmountOut;
     }
@@ -235,37 +223,37 @@ contract BMath is BConst, BNum {
     // eF = exitFee                                                                              //
     **********************************************************************************************/
     function calcPoolInGivenSingleOut(
-        uint tokenBalanceOut,
-        uint tokenWeightOut,
-        uint poolSupply,
-        uint totalWeight,
-        uint tokenAmountOut,
-        uint swapFee
-    )
-        public pure
-        returns (uint poolAmountIn)
-    {
-
+        uint256 tokenBalanceOut,
+        uint256 tokenWeightOut,
+        uint256 poolSupply,
+        uint256 totalWeight,
+        uint256 tokenAmountOut,
+        uint256 swapFee
+    ) public pure returns (uint256 poolAmountIn) {
         // charge swap fee on the output token side
-        uint normalizedWeight = bdiv(tokenWeightOut, totalWeight);
+        uint256 normalizedWeight = bdiv(tokenWeightOut, totalWeight);
         //uint tAoBeforeSwapFee = tAo / (1 - (1-weightTo) * swapFee) ;
-        uint zoo = bsub(BONE, normalizedWeight);
-        uint zar = bmul(zoo, swapFee);
-        uint tokenAmountOutBeforeSwapFee = bdiv(tokenAmountOut, bsub(BONE, zar));
+        uint256 zoo = bsub(BONE, normalizedWeight);
+        uint256 zar = bmul(zoo, swapFee);
+        uint256 tokenAmountOutBeforeSwapFee = bdiv(
+            tokenAmountOut,
+            bsub(BONE, zar)
+        );
 
-        uint newTokenBalanceOut = bsub(tokenBalanceOut, tokenAmountOutBeforeSwapFee);
-        uint tokenOutRatio = bdiv(newTokenBalanceOut, tokenBalanceOut);
+        uint256 newTokenBalanceOut = bsub(
+            tokenBalanceOut,
+            tokenAmountOutBeforeSwapFee
+        );
+        uint256 tokenOutRatio = bdiv(newTokenBalanceOut, tokenBalanceOut);
 
         //uint newPoolSupply = (ratioTo ^ weightTo) * poolSupply;
-        uint poolRatio = bpow(tokenOutRatio, normalizedWeight);
-        uint newPoolSupply = bmul(poolRatio, poolSupply);
-        uint poolAmountInAfterExitFee = bsub(poolSupply, newPoolSupply);
+        uint256 poolRatio = bpow(tokenOutRatio, normalizedWeight);
+        uint256 newPoolSupply = bmul(poolRatio, poolSupply);
+        uint256 poolAmountInAfterExitFee = bsub(poolSupply, newPoolSupply);
 
         // charge exit fee on the pool token side
         // pAi = pAiAfterExitFee/(1-exitFee)
         poolAmountIn = bdiv(poolAmountInAfterExitFee, bsub(BONE, EXIT_FEE));
         return poolAmountIn;
     }
-
-
 }
